requirements.txt
requests==2.7.0
PyYAML==3.11

docutils>=0.10
Sphinx==1.1.3
nose==1.3.0
tox==1.4
colorama>=0.2.5,<=0.3.3
mock==1.0.1
wheel==0.24.0
